third_party/WebKit/LayoutTests/typedcssom/stylevalue-subclasses/keywordvalue-objects/constructor.html

<!doctype html>
<meta charset="utf-8">
<title>CSSKeywordValue constructor tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-csskeywordvalue-csskeywordvalue">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
'use strict';

const gValidTestValues = [
  { value: 'initial', desc: 'a CSS keyword' },
  { value: 'customLemon', desc: 'an unsupported CSS keyword' },
  { value: '3! + 4@', desc: 'a string containing multiple tokens' },
  { value: '☺', desc: 'a unicode string' },
  { value: [3], desc: 'an object that serializes to a nonempty string' }
];

for (const {value, desc} of gValidTestValues) {
  test(() => {
    const result = new CSSKeywordValue(value);
    assert_equals(result.value, value.toString());
  }, 'CSSKeywordValue can be constructed from ' + desc);
}

const gInvalidTestValues = [
  { value: '', desc: 'an empty string' },
  { value: [], desc: 'an object that serializes to empty string' },
];

for (const {value, desc} of gInvalidTestValues) {
  test(() => {
    assert_throws(new TypeError(), () => new CSSKeywordValue(value));
  }, 'Constructing CSSKeywordValue with ' + desc + ' throws a TypeError');
}

</script>
